src/functions/notify/requirements.txt
azure-functions==1.21.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
idna==3.10
iniconfig==2.0.0
jwt==1.3.1
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
urllib3==2.2.3
